feat(sidebar): inline rename for account rows
Adds editingRowId state to SidebarView, a Rename context-menu item,
and double-click-to-rename on account rows in both Current and
Investments sections. Stores call AccountStore.rename. UITestIdentifier
added so the future group flow can target the same menu item.

Extracted SidebarView+Sections.swift to keep SidebarView.swift under
the file_length SwiftLint limit after the new rename wiring added lines.
